last update

+++ app/Http/Controllers/ListingController.php
@@ -53,6 +53,9 @@ class ListingController extends Controller
                 ->store('logos', 'public');
         }
 
+        $formFields['user_id'] = auth()
+            ->id();
+
         Listing::create($formFields);
 
         return redirect('/')
@@ -69,6 +72,13 @@ class ListingController extends Controller
     //update listing
     public function update(Request $request, Listing $listing)
     {
+        // Make sure logged-in user is the owner
+        if ($listing->user_id != auth()->id())
+        {
+            abort(403, 'Unauthorized Action');
+        }
+
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => 'required',
@@ -96,9 +106,24 @@ class ListingController extends Controller
     //delete listing
     public function destroy(Listing $listing)
     {
+        if ($listing->user_id != auth()->id())
+        {
+            abort(403, 'Unauthorized Action');
+        }
+
         $listing->delete();
         return redirect('/')
             ->with('message',
         'Listing deleted successfully!');
+    }
+
+    //manage listings
+    public function manage()
+    {
+        return view('listings.manage',
+            ['listings' => auth()
+            ->user()
+            ->listings()
+            ->get()]);
     }
 }
